Choosing deploy route that requires pre-rendered _site.
If it weren't for the R code in here, there is a quarto plugin
for netlify that would allow netlify to build the site itself.

But no such R plugin. SO, need to pre-render. The downside is
that we need to version control site/ (gross). But upside is
that should no longer need manual quarto publish netlify
command.

The alternative, perhaps for a later date is to have GH actions
do the build (and maybe the deploy and host, too?) while netlify
only deploys and hosts and manages the domain. Not sure which
exactly does the deploy/host, but don't care rn.